fix signed/unsigned warnings

+++ openudm/CellularModel.cpp
@@ -76,7 +76,7 @@ void CellularModel::LoadCurrentPopulation(std::string popData, int curPopColumn,
 	ExtractCSV(popData, numColumns, curPopColumn, tempCP);
 
 	//convert string data to double
-	for (int w = 0; w != numWards; ++w) {
+	for (size_t w = 0; w != numWards; ++w) {
 		curPop[w] = GetDoubleFromString(tempCP[w]);
 	}
 }
@@ -87,7 +87,7 @@ void CellularModel::LoadFuturePopulation(std::string popData, int futPopColumn, 
 	ExtractCSV(popData, numColumns, futPopColumn, tempFP);
 
 	//convert string data to double
-	for (int w = 0; w != numWards; ++w) {
+	for (size_t w = 0; w != numWards; ++w) {
 		futPop[w] = GetDoubleFromString(tempFP[w]);
 	}
 }
@@ -98,7 +98,7 @@ void CellularModel::LoadWardDensity(std::string densityData, int densityColumn, 
 	ExtractCSV(densityData, numColumns, densityColumn, tempWardStr);
 
 	//convert string data to double
-	for (int w = 0; w != numWards; ++w) {
+	for (size_t w = 0; w != numWards; ++w) {
 		wardDensity[w] = GetDoubleFromString(tempWardStr[w]);
 	}
 
@@ -192,7 +192,7 @@ void CellularModel::RunModel() {
 void CellularModel::CreateWards() {
 
 	//create wards
-	for (int w = 0; w != numWards; ++w) {
+	for (size_t w = 0; w != numWards; ++w) {
 		wards.push_back(new UDMWard());
 	}	
 }
@@ -202,8 +202,8 @@ void CellularModel::AssignWardCells() {
 	//ASSIGN RASTER CELLS TO WARDS 		
 
 	//--read pixels into allocated wards using wardID
-	for (int r = 0; r != rasterRows; ++r) {
-		for (int c = 0; c != rasterCols; ++c) {
+	for (size_t r = 0; r != rasterRows; ++r) {
+		for (size_t c = 0; c != rasterCols; ++c) {
 			if (iWardID.data[r][c] != iWardID.NODATA_value) {
 				wards[iWardID.data[r][c]]->cells.push_back(new UDMCell(c, r));
 			}			
@@ -216,7 +216,7 @@ void CellularModel::CalculatePopulationChange() {
 	//CALCULATE WARD POPULATION CHANGE 
 
 	//calculate population change for all wards and flag devReq as false where popChange !> 0
-	for (int w = 0; w != numWards; ++w) {
+	for (size_t w = 0; w != numWards; ++w) {
 		wards[w]->popChange = futPop[w] - curPop[w];
 
 		if (wards[w]->popChange <= 0.0) {
@@ -228,10 +228,10 @@ void CellularModel::CalculatePopulationChange() {
 void CellularModel::AssignZones() {
 
 	//assign zones
-	for (int w = 0; w != numWards; ++w) {	//for all wards
+	for (size_t w = 0; w != numWards; ++w) {	//for all wards
 		//if (wards[w]->devReq) {				//which require development
 			if (!wards[w]->cells.empty()) {		//if ward has some cells
-				for (int c = 0; c != wards[w]->cells.size(); ++c) {	//for all cells within ward
+				for (size_t c = 0; c != wards[w]->cells.size(); ++c) {	//for all cells within ward
 
 					int zid = zoneID.data[wards[w]->cells[c]->row][wards[w]->cells[c]->col];	//read zoneID 
 
@@ -242,7 +242,7 @@ void CellularModel::AssignZones() {
 							bool foundZ = false;
 
 							//find out if we already have a zone with this ID
-							for (int z = 0; z != wards[w]->zones.size(); ++z) {
+							for (size_t z = 0; z != wards[w]->zones.size(); ++z) {
 								if (wards[w]->zones[z]->ID == zid) {
 									//add cell to zone[z] and flag found
 									wards[w]->zones[z]->cells.push_back(wards[w]->cells[c]);
@@ -304,10 +304,10 @@ void CellularModel::CalculateRequiredDevelopment() {
 	//pop density per cell and pop density (area)
 	//required cells for new development
 
-	for (int w = 0; w != numWards; ++w) {														//all wards
+	for (size_t w = 0; w != static_cast<size_t>(numWards); ++w) {														//all wards
 		//if (wards[w]->devReq) {
 			int devCells = 0;																	//reset for each ward
-			for (int c = 0; c != wards[w]->cells.size(); ++c) {									//all cells in ward
+			for (size_t c = 0; c != wards[w]->cells.size(); ++c) {									//all cells in ward
 				int dev = devLand.data[wards[w]->cells[c]->row][wards[w]->cells[c]->col];
 				if (dev == 1) {
 					++devCells;
@@ -319,7 +319,7 @@ void CellularModel::CalculateRequiredDevelopment() {
 
 			if (densityProvided) {
 
-				for (int w = 0; w != numWards; ++w) {
+				for (size_t w = 0; w != numWards; ++w) {
 					wards[w]->density = wardDensity[w];
 					wards[w]->cellDensity = wardDensity[w] * devRes * devRes;
 				}
@@ -341,10 +341,10 @@ void CellularModel::FindOverflowWards() {
 	//FIND OVERFLOW WARDS 
 
 	//find out if we have any overflow wards
-	for (int w = 0; w != numWards; ++w) {
+	for (size_t w = 0; w != numWards; ++w) {
 		if (wards[w]->devReq) {
 			int suitCells = 0;											//reset for each ward iteration 
-			for (int z = 0; z != wards[w]->zones.size(); ++z) {
+			for (size_t z = 0; z != wards[w]->zones.size(); ++z) {
 				suitCells += wards[w]->zones[z]->cells.size();
 			}
 			wards[w]->suitDevCells = suitCells;
@@ -360,7 +360,7 @@ void CellularModel::DevelopNonOverflowWards() {
 	//DEVELOP NON-OVERFLOW WARDS BEGIN--------	
 
 	//sort all zones in ward by highest avgSuit
-	for (int w = 0; w != numWards; ++w) {
+	for (size_t w = 0; w != numWards; ++w) {
 		if (wards[w]->devReq) {
 			if (!wards[w]->overflow) {
 				std::sort(wards[w]->zones.begin(), wards[w]->zones.end(), HighestZoneAVG);
@@ -369,20 +369,20 @@ void CellularModel::DevelopNonOverflowWards() {
 	}
 
 	//develop cells in non-overflow wards
-	for (int w = 0; w != numWards; ++w) {
+	for (size_t w = 0; w != numWards; ++w) {
 		if (wards[w]->devReq) {
 			if (!wards[w]->overflow) {
 
 				int devCells = 0;	//keep track of how many cells have been developed - reset for each ward
 
-				for (int z = 0; z != wards[w]->zones.size(); ++z) {		//all zones
+				for (size_t z = 0; z != wards[w]->zones.size(); ++z) {		//all zones
 					if (devCells < wards[w]->reqDevCells) {				//requiring further development 
 
 						//find out if we're in the spread zone (if there is one... <=)
-						if ((wards[w]->zones[z]->cells.size() + devCells) <= wards[w]->reqDevCells) {	//not in spread zone
+						if ((static_cast<int>(wards[w]->zones[z]->cells.size()) + devCells) <= wards[w]->reqDevCells) {	//not in spread zone
 
 							//develop all cells in zone - keep track of devCells
-							for (int c = 0; c != wards[w]->zones[z]->cells.size(); ++c) {
+							for (size_t c = 0; c != wards[w]->zones[z]->cells.size(); ++c) {
 								wards[w]->zones[z]->cells[c]->devStatus = true;
 								++devCells;
 								
@@ -392,7 +392,7 @@ void CellularModel::DevelopNonOverflowWards() {
 						else {	//spread zone - seed and spread development from most suitable cell
 
 							//read cell suitability from raster
-							for (int c = 0; c != wards[w]->zones[z]->cells.size(); ++c) {
+							for (size_t c = 0; c != wards[w]->zones[z]->cells.size(); ++c) {
 								wards[w]->zones[z]->cells[c]->suit =
 									cellSuit.data[wards[w]->zones[z]->cells[c]->row][wards[w]->zones[z]->cells[c]->col];
 							}
@@ -421,7 +421,7 @@ void CellularModel::DevelopNonOverflowWards() {
 								nbrs.clear();
 
 								//get neighbours
-								for (int c = 0; c != wards[w]->zones[z]->cells.size(); ++c) {
+								for (size_t c = 0; c != wards[w]->zones[z]->cells.size(); ++c) {
 									int cDiff = (seeds[seeds.size() - 1]->col + 1) - (wards[w]->zones[z]->cells[c]->col + 1);
 									int rDiff = (seeds[seeds.size() - 1]->row + 1) - (wards[w]->zones[z]->cells[c]->row + 1);
 
@@ -444,7 +444,7 @@ void CellularModel::DevelopNonOverflowWards() {
 								}
 								
 								//develop neighbouring cells
-								for (int n = 0; n != nbrs.size(); ++n) {
+								for (size_t n = 0; n != nbrs.size(); ++n) {
 									if (!nbrs[n]->devStatus) {						//not already developed										
 										if (devCells < wards[w]->reqDevCells) {
 											nbrs[n]->devStatus = true;
@@ -512,7 +512,7 @@ void CellularModel::WriteOverflowWards(std::string overflowData) {
 
 		opfile << "ReqPopCellDensityInWardCells" << "\n";
 
-		for (int w = 0; w != wards.size(); ++w) {
+		for (size_t w = 0; w != wards.size(); ++w) {
 
 			//ward label
 			opfile << wardLabel[w] << ",";
@@ -588,14 +588,14 @@ void CellularModel::DevelopOverflowWards() {
 		}
 	}*/
 
-	for (int w = 0; w != numWards; ++w) {
+	for (size_t w = 0; w != numWards; ++w) {
 
 		int devCells = 0;	//keep track of how many cells have been developed - reset for each ward
 
 		if (wards[w]->devReq) {
 			if (wards[w]->overflow) {
-				for (int z = 0; z != wards[w]->zones.size(); ++z) {
-					for (int c = 0; c != wards[w]->zones[z]->cells.size(); ++c) {
+				for (size_t z = 0; z != wards[w]->zones.size(); ++z) {
+					for (size_t c = 0; c != wards[w]->zones[z]->cells.size(); ++c) {
 						wards[w]->zones[z]->cells[c]->devStatus = true;
 						++devCells;
 						wards[w]->devCells = devCells;			//keep track of how many cells have been developed	
@@ -609,8 +609,8 @@ void CellularModel::DevelopOverflowWards() {
 void CellularModel::SetCurrentDev() {
 
 	//set currently developed land 
-	for (int r = 0; r != rasterRows; ++r) {
-		for (int c = 0; c != rasterCols; ++c) {
+	for (size_t r = 0; r != rasterRows; ++r) {
+		for (size_t c = 0; c != rasterCols; ++c) {
 			if (devLand.data[r][c] == 1) {
 				finalDev.data[r][c] = 1;
 			}
@@ -621,8 +621,8 @@ void CellularModel::SetCurrentDev() {
 void CellularModel::SetNoData() {
 
 	//create NODATA mask using iWardID raster
-	for (int r = 0; r != rasterRows; ++r) {
-		for (int c = 0; c != rasterCols; ++c) {
+	for (size_t r = 0; r != rasterRows; ++r) {
+		for (size_t c = 0; c != rasterCols; ++c) {
 			if (iWardID.data[r][c] == iWardID.NODATA_value) {
 				finalDev.data[r][c] = finalDev.NODATA_value;
 			}
@@ -633,11 +633,11 @@ void CellularModel::SetNoData() {
 void CellularModel::SetFutureDev() {
 
 	//set future developed land
-	for (int w = 0; w != wards.size(); ++w) {
+	for (size_t w = 0; w != wards.size(); ++w) {
 		if (wards[w]->devReq) {
 			if (!wards[w]->zones.empty()) {
-				for (int z = 0; z != wards[w]->zones.size(); ++z) {
-					for (int c = 0; c != wards[w]->zones[z]->cells.size(); ++c) {
+				for (size_t z = 0; z != wards[w]->zones.size(); ++z) {
+					for (size_t c = 0; c != wards[w]->zones[z]->cells.size(); ++c) {
 
 						if (wards[w]->zones[z]->cells[c]->devStatus) {
 							finalDev.data[wards[w]->zones[z]->cells[c]->row][wards[w]->zones[z]->cells[c]->col] = 2;
